Move versions test
